Some cleanup in the Employee entity class.

+++ src/main/java/com/example/companydbapi/repository/EmployeeReporitory.java
@@ -1,8 +1,8 @@
 package com.example.companydbapi.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.example.companydbapi.model.Employee;
 
-public interface EmployeeReporitory extends JpaRepository<Employee, Integer> {
+public interface EmployeeReporitory extends JpaRepository<Employee, Long> {
+	
 }
